Fix crash when starting from non ASCII path
This fixes a crash that occurs when the launcher is started from a
folder whose absolute path does contain non ASCII characters. This
was due to the use of the WshRuntimeLibrary (a wrapper around COM
libraries) that does not handle Unicode (as far as my deductions
go). When assigning the target path to the shortcut object, the
runtime would error.

This commit replaces the WshRuntimeLibrary with a COM wrapper around
the IShellLink interface, using its Unicode oriented functions.

The WshRuntimeLibrary is removed from dependencies and the .NET
framework is updated to 4.8 as 4.0 has reached EOL.

Tested on: Win10 64-bits.